requirements.txt
python-telegram-bot[all]>=21.0
web3>=6.0.0
aiosqlite>=0.19.0
cryptography>=41.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
